+++ pages/public-docspace/js/index.js
@@ -17,63 +17,52 @@
 			'onlyoffice-docspace-block'
 		);
 		const oodspConfigs = [];
+		const oodspPublicConfigs = [];
 
 		for ( const frame of frames ) {
-			oodspConfigs.push( JSON.parse( frame.dataset.config ) );
+			let config = JSON.parse( frame.dataset.config );
+			config = Object.assign( config, defaultConfig );
+
+			if ( config.requestToken && config.requestToken.length > 0 ) {
+				oodspPublicConfigs.push( config );
+			} else {
+				oodspConfigs.push( config );
+			}
 		}
 
 		const countElements = oodspConfigs.length;
-
-		if ( countElements === 0 ) {
-			return;
-		}
 
 		DocspaceIntegrationSdk.initScript(
 			'oodsp-api-js',
 			_oodspDocspacePublic.docspaceUrl
 		)
 			.then( function () {
-				for ( let i = 0; i < countElements; i++ ) {
-					oodspConfigs[ i ] = Object.assign(
-						oodspConfigs[ i ],
-						defaultConfig
-					);
+				_initFrames( oodspPublicConfigs );
 
+				for ( let i = 0; i < countElements; i++ ) {
 					if ( i === 0 ) {
 						if (
 							_oodspDocspacePublic.isAnonymous ||
 							! _oodspDocspacePublic.docspaceUser
 						) {
-							DocspaceIntegrationSdk.logout(
-								oodspConfigs[ 0 ].frameId,
-								function () {
-									_initAllFrames( oodspConfigs, true );
-								}
-							);
-						} else {
-							DocspaceIntegrationSdk.loginByPasswordHash(
-								oodspConfigs[ 0 ].frameId,
-								_oodspDocspacePublic.docspaceUser.user_name,
-								function () {
-									return _oodspDocspacePublic.docspaceUser
-										.password_hash;
-								},
-								function () {
-									_initAllFrames( oodspConfigs, false );
-								},
-								function () {
-									DocspaceIntegrationSdk.logout(
-										oodspConfigs[ 0 ].frameId,
-										function () {
-											_initAllFrames(
-												oodspConfigs,
-												true
-											);
-										}
-									);
-								}
-							);
+							_showUnauthorizedTemplates( oodspConfigs );
+							return;
 						}
+
+						DocspaceIntegrationSdk.loginByPasswordHash(
+							oodspConfigs[ 0 ].frameId,
+							_oodspDocspacePublic.docspaceUser.user_name,
+							function () {
+								return _oodspDocspacePublic.docspaceUser
+									.password_hash;
+							},
+							function () {
+								_initFrames( oodspConfigs );
+							},
+							function () {
+								_showUnauthorizedTemplates( oodspConfigs );
+							}
+						);
 					} else {
 						DocSpace.SDK.initSystem( {
 							frameId: oodspConfigs[ i ].frameId,
@@ -89,31 +78,30 @@
 				// eslint-disable-next-line no-console
 				console.error( error );
 
-				for ( const config of oodspConfigs ) {
+				for ( const config of [
+					...oodspConfigs,
+					...oodspPublicConfigs,
+				] ) {
 					oodsp.templates.docspaceUnavailable( config.frameId );
 				}
 			} );
 	} );
 
-	const _initAllFrames = ( oodspConfigs, requiredRequestToken ) => {
+	const _initFrames = ( oodspConfigs ) => {
 		for ( const config of oodspConfigs ) {
-			if (
-				requiredRequestToken &&
-				( ! config.hasOwnProperty( 'requestToken' ) ||
-					config.requestToken.length <= 0 )
-			) {
-				if ( DocSpace.SDK.frames[ config.frameId ] !== null ) {
-					DocSpace.SDK.frames[ config.frameId ].destroyFrame();
-				}
-
-				oodsp.templates.docspaceUnauthorized( config.frameId );
-
-				continue;
-			}
-
 			config.src = DocSpace.SDK.src;
 
-			DocSpace.SDK.frames[ config.frameId ].initFrame( config );
+			DocSpace.SDK.initFrame( config );
+		}
+	};
+
+	const _showUnauthorizedTemplates = ( oodspConfigs ) => {
+		for ( const config of oodspConfigs ) {
+			if ( DocSpace.SDK.frames[ config.frameId ] ) {
+				DocSpace.SDK.frames[ config.frameId ].destroyFrame();
+			}
+
+			oodsp.templates.docspaceUnauthorized( config.frameId );
 		}
 	};
 } )( jQuery );